complete: Learn type=module, node(http-server), force refresh
Adding type=module to script tag causes the js file to be treated as a
module. Meaning that all functions and variables are not in global scope
and hence cannot be accessed.

Due to a CROS error (browser could not access file/// extension in a
safe manner), I had to use nodejs to create a server. I used http-server
in the directory I was working in.

The changes made in my js file were not being reflected in the browser.
I had to force refresh by ctrl + shift + r. This caused the cache to be
cleared which was the cause of the issue.